\name{sugar_duplicated}
\alias{sugar_duplicated}
\title{Rcpp Sugar -- duplicated}
\usage{
  sugar_duplicated(x, ...)
}
\arguments{
  \item{x}{a vector of storage mode: double, integer,
  character}

  \item{...}{additional arguments used for compatibility}
}
\description{
  This function implements the Rcpp sugar function
  \code{duplicated}.
}

