\name{sugar_table}
\alias{sugar_table}
\title{Rcpp Sugar -- table}
\usage{
  sugar_table(x, ...)
}
\arguments{
  \item{x}{a vector of storage mode: character, double,
  integer}

  \item{...}{additional arguments used for compatibility}
}
\description{
  This function implements the Rcpp sugar function
  \code{table}.
}

