\name{sugar_unique}
\alias{sugar_unique}
\title{Rcpp Sugar -- unique}
\usage{
  sugar_unique(x, ...)
}
\arguments{
  \item{x}{a vector of storage mode: double, integer,
  character}

  \item{...}{additional arguments used for compatibility}
}
\description{
  This function implements the Rcpp sugar function
  \code{unique}.
}

