\name{truncate}
\alias{truncate}
\title{Rcpp Sugar -- truncate}
\usage{
  truncate(x, n = 80)
}
\arguments{
  \item{x}{a vector of strings}

  \item{n}{an integer; length of substrings}
}
\description{
  This function truncates a vector of strings \code{x} up
  to a width \code{n}.
}

