\name{collapse}
\alias{collapse}
\title{Rcpp Sugar -- collapse}
\usage{
  collapse(x, collapse = "")
}
\arguments{
  \item{x}{a vector of storage mode: character}

  \item{collapse}{a string (character vector of length 1)
  to collapse over}
}
\description{
  This function implements the Rcpp sugar function
  \code{collapse}, with augmentation to handle a custom
  \code{collapse} argument.
}

