\name{match}
\alias{match}
\title{Rcpp Sugar -- match}
\usage{
  match(x, y, ...)
}
\arguments{
  \item{x}{a vector of storage mode: character, double,
  integer}

  \item{y}{a vector of storage mode: character, double,
  integer}

  \item{...}{additional arguments used for compatibility}
}
\description{
  This function implements the Rcpp sugar function
  \code{match}.
}

